added seeding for unspecified category and updated delete to reassing expenses'categories to unspecified with the unability of deleting the unspecified category

+++ Services/CategoryService.cs
@@ -9,19 +9,22 @@ namespace ExpenseTrackerAPI.Services;
 public class CategoryService : ICategoryService
 {
     private readonly ICategoryRepository _categoryRepository;
+
+    private readonly IExpenseRepository _expenseRepository;
     private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public CategoryService(ICategoryRepository categoryRepository, IHttpContextAccessor httpContextAccessor)
+    public CategoryService(ICategoryRepository categoryRepository, IExpenseRepository expenseRepository, IHttpContextAccessor httpContextAccessor)
     {
         _categoryRepository = categoryRepository;
         _httpContextAccessor = httpContextAccessor;
+        _expenseRepository = expenseRepository;
     }
 
     public async Task<CategoryDto?> CreateCategoryAsync(CategoryCreateDto dto)
     {
         var newCategory = dto.ToEntity();
-        bool existingCategory = await _categoryRepository.GetCategoryExistsByNameAsync(newCategory.Name);
-        if (existingCategory)
+        var existingCategory = await _categoryRepository.GetCategoryByNameAsync(newCategory.Name);
+        if (existingCategory != null)
         {
             return null;
         }
@@ -76,6 +79,10 @@ public class CategoryService : ICategoryService
         var existingCategory = await _categoryRepository.GetCategoryAsync(id);
         if (existingCategory is null)
             return false;
+        if (existingCategory.Name == "Unspecified")
+            throw new InvalidOperationException("Cannot delete the 'Unspecified' category.");
+        var unspecified = await _categoryRepository.GetCategoryByNameAsync("Unspecified");
+        await _expenseRepository.ReassignCategoryAsync(id, unspecified!.Id);
         await _categoryRepository.DeleteCategoryAsync(existingCategory);
         return true;
     }
